+++ wolox-books/src/app/screens/User/screens/Sign-up/sign-up.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { FormGroup, FormControl } from '@angular/forms';
-import { UserService } from './user.service';
+import { UserService } from '../../user.service';
 import { Router } from '@angular/router';
